Fix CC20-MT payload keystream alignment
The payload does not start immediately at the beginning of the src (or
dest) buffers. The payload length comes first, with a width of 'aadlen'
(i.e., four) bytes. This gets encrypted using a different keystream, so
the beginning of the payload keystream still gets read from index zero,
but it needs to be XORed against src with an offset of 'aadlen' (and
written into dest with the same offset).