Correct ccmt_crypt() missing-MT-context behavior
If ccmt_crypt() is passed a chachapoly context that does not have an
embedded MT context (such as if it were initialized as a serial cipher),
don't return 0, since that would indicate that the encryption/decryption
was successful. There are at least three options worth considering. For
the first option, we could return immediately with an error condition.
For the second option, we could transparently fall back to the serial
implementation. For the third option, chosen in this commit, we could
initialize an MT context so that future calls to ccmt_crypt() are more
useful.