Significantly refine ChaCha20 worker processes
Implement an "err" function in the ChaCha20 worker processes, which can
be used to write formatted messages to stderr if DEBUGMODE is enabled,
but uses the preprocessor to skip them entirely when DEBUGMODE is
disabled.

Implement a function for printing strings of hexadecimal data to
standard error.

Replace calls to fread() with calls to read().

Refactor readInt() and readBytes() helper functions to use a
readBinLoop() helper function instead, to better tolerate partial reads
of the pipe.