Add chachapoly worker to build system
Update Makefile.in to compile the external chacha20/poly1305 worker
process along with the other helper binaries.

Rename the worker process to "(hpn)ssh-chachapoly-helper", and rename
the source file to "ssh-ccp-helper.c".

Add a preprocessor define variable for the path to the external worker
process, modeled after the approach used for other helper binaries.