Add an option to reduce cipher-bench packet size
Add the "-p" run-time option to cipher-bench, which sets the packet size
to something smaller than the default (which was the maximum,
SSH_IOBUFSZ).

Also add bounds-checking for both "-p" and "-n", such that "-p" triggers
an error for any negative packet sizes, and "-n" triggers an error for
any packet counts smaller than 1. An excessively large "-p" will trigger
a warning and reduce the packet size to the maximum, SSH_IOBUFSZ. No
upper-bound is enforced for "-n", since the limiting factor is the
host's memory.