Fix a failure to join a thread in CC20-MT
Add a pthread_join() call to the cipher_free() codepath in the
multithreaded implementation of chacha20-poly1305. A recent commit
implemented a helper thread to asynchronously manage the worker threads
when the main thread transitions to a new set of keystreams, reducing
the likelihood that the main thread will block while assigning task to
the worker threads. In rare circumstances, it was possible for the
cipher context to be freed before this helper thread returns,
potentially leading to writes to regions of memory that had already been
freed (and potentially reallocated), especially in cases where the
rekeying frequency is high and the number of CPU cores is low.