First pass at integrating ISSHD with HPN.
The main area that concerns me at the moment is caused by
the sshd split. NERSC_MOD creates a number of externs that will
be used by both sshd and sshd-session. So I'm concerned that there
may be issues with overwrites, bad data, etc, happening.

I need to understand more about this to really understand
if this is an issue and, if so, how to fix it. That said, it compiles and
the client runs. I'll be starting regression tests shortly.