Start thinking about using xxHash instead of blake512 for resume function.
xxHash is not a cryptographic hash *but* we shouldn't need one to ensure that
the file fragments align. So we can make use of the enhanced speed and reduce
the dependency on a crypto library to support the resume function. So far this
commit is only a comment in scp.c to act as a placeholder