Merged the HPN-SSH 18.7 port for the runup to the OpenSSH 10.0 release. Added regression test.
This version of the MPTCP code gets rid of the configure test and
relies on the socket not being created if they request mptcp
and it doesn't exist on their system.

Additionally, created a regression test for mptcp based on
the simple connect test. If this test fails it will *not*
exit the tests as a whole but will issue a warning instead.
This isn't properly communicated to the master make file but
it's what we have. If a warning is thrown we do keep a copy of the
logs.