Add RFC 8305 compliance to HPN-SSH.
This is a patched based on one written by
Kim Minh Kaplan and submitted to OpenSSH in 2018.
This implements *some* features of RFC 8305 but,
significnatly, does not do asynchronous DNS queries.
It is currently only implemented for direct and proxy
connections as found in sshconnect.c. This is the first
version of the patch and is not in its final form. Eventually
the RFC 8305 methods will be moved their own file and the
necessary methods made public for use elsewhere in the
code base.