Happy Eyeballs was being hammered by IPQoS defaults
The default IPQoS for SSH had changed from
interactive = lowdelay
bulk = throughput
to
interactive = af21
bulk = cs1

The result of this being that in test scenarios, where we used
tc qdisc to insert delay on IPv6, the IPv4 connections ended up
having the same effective delay as the IPv6. This meant that we'd
always end up using the IPv6 connection even though IPv4 was
not being delayed. I don't really understand DSCP enough to undertand
why this is happening but this might be a test artifact. In any case,
being that both IPv4 and IPv6 had the same effective delay the IPv6
connection would always end up winning the race because of the
250ms pause between the IPv6 and IPv4 attempts.

This commit stil lhas too many debug statements in it but
it works for testing.