Testing for fips_enabled in serv|readconf.c didn't work.
For some reason using fips enabled to set options.disable_multithreaded
was causing errors on the server side. I don't know why. So I moved the
test to where we call cipher_switch() for the parallel aes cipher.

I don't think I'm going to put in any overrides for the parallel
CC20 cipher as CC20, in any form, is not FIPS compliant and it's
up to the admin to ensure that it's not being offered as an option.
That's part of the default on RH systems.