Enabled Valgrind rekey tests aka valgrind-5.
There was a problem in how valgrind was interacting with the
aes-ctr-mt cipher. The method of serialization was causing hangs
during the rekey regression test. By using the --fair-sched=try
options valgrind will use futexes for locking which is less likely
to cause a race. This locking was most prominent on 2 core VM like those used
in the CI tests on GitHub.