Disable the parallel AES-CTR cipher under FIPS mode.
While the AES-CTR algorithm is FIPS certified this implementation
is not. This causes FIPS to automatically error out when the aes-ctr-mt
cipher is loaded. This can be resolved by using the DisableMTAES options but
this requires the user to be aware of it and know why FIPS is erroring out.
This will automatically detect if FIPS mode is enabled on fedora variants that
write to /proc/sys/crypto/fips_enabled. In the event this entry doesn't exist
or that we can't read from it we assume that FIPS mode is not enabled and
we allow parallel ciphers.

This does not apply to the ChaCha20 cipher as the algorithm itself is not
FIPS validated and it's up to the admin to make sure it's not being
offered as an option.