Various fixes to resolve problem with the CI tests
These fixes include pushing the SCP resume tests to their
own test script. They don't work under windows so we now
skip them.

It also includes a change to the XOR function in cipher-ctr-mt.
We were trying to do a 64 bit XOR but it was causing UBs in
some of the CI tests. Until we have that resolved we need to
stick with 32 bit XORs. Which we *really* need to resolve.

The last major thing was that valgrind was consistently hanging on
the rekey test (valgrind-5). Using the --fair-sched=try option
for thread serialization in valgrind resolves that issue.