Updated HPN-README. Changed fips_enabled().
fips_enabled() in misc.c originally had debug messages that
reported if the parallel ciphers were enabled or not. This
was the wrong place for those messages. They should be in
session.c and ssh.c when we check the FIPS status.

Also, the function had unclear variable names. These were
updated to be more descriptive.