mptcp: do not always define IPPROTO_MPTCP
IPPROTO_MPTCP has been introduced in the Linux kernel v5.6, more than 5
years ago. It has been added in netinet the same year. In other words,
it should be available on systems having a kernel supporting MPTCP.

In this case, IPPROTO_MPTCP can be used to identify OS where MPTCP is
supported directly where it is used. This will simplify the next patch.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>