sshd: use MPTCP by default if supported
If not, a "plain" TCP socket is used like before.

It is interesting to do so to let only the clients deciding if MPTCP is
going to be used or not: clients are probably the ones who know if they
can benefit from using MPTCP or not, not the server side. Like that, it
is easier to do experimentations with MPTCP without having to explicitly
ask sysadmin to enable this option with a long explanation.

On the server side, it is safe to enable MPTCP by default because MPTCP
will only be used if requested by clients. If not, a "plain" TCP socket
is returned to the application, not impacting the performances and then
not introducing any regressions. Note that GO server applications are
using MPTCP by default since v1.24 [1].

Link: https://go.dev/doc/go1.24#netpkgnet [1]
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>