configure: remove MPTCP check
It is not used, and not recommended: when packages are distributed, a
different kernel will be used at build time and at run time. These
kernels can have different versions, different parameters, etc. Also, on
Linux, it is common to use containers at build time to reproduce a
specific environments, except that it doesn't include the right kernel.

Instead, IPPROTO_MPTCP can be checked to see if the OS supports it or
not.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>